+++ package.json
@@ -24,7 +24,7 @@
     "next": "12.3.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.41.0"
+    "react-hook-form": "^7.41.5"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.20.2",
