+++ src/containers/Profile/Profile.error.tsx
@@ -1,0 +1,10 @@
+import React from "react"
+
+// todo error boundary
+type Props = {}
+
+const ProfileLoading: React.FC<Props> = ({}) => {
+  return <div className="profile">에러발생 이멀전시</div>
+}
+
+export default ProfileLoading
