+++ src/containers/Profile/Profile.loading.tsx
@@ -1,0 +1,9 @@
+import React from "react"
+
+type Props = {}
+
+const ProfileLoading: React.FC<Props> = ({}) => {
+  return <div className="profile">loading</div>
+}
+
+export default ProfileLoading
