+++ src/libs/api/users.ts
@@ -1,4 +1,5 @@
 import Repository from "./repository"
+import coqualityAxiosClient from "./client"
 
 interface IUser {
   nickname: string
@@ -10,6 +11,31 @@ interface IUser {
 //   return x
 // }
 
+interface IGetPostByIdParams {
+  userId?: number
+}
+
+export type PostSortType = "LATEST" | "VIEWS"
+
+export interface IPostType {
+  id: number
+  userId: number
+  title: string 
+  contents: string
+  thumbnail: null
+  primaryCategory: string
+  postStatusCode: string
+  summary: string
+  views: number
+  commentCount: number
+  createdAt: string
+}
+
+export interface ProfileModifyType {
+  email: string,
+  nickname: string,
+  userSummary: string
+}
 export class UsersRepository extends Repository {
   public async readMyInfo(): Promise<IUser> {
     const response = await this.client.get(`/users/read`, {
@@ -36,3 +62,9 @@ export class UsersRepository extends Repository {
     )
   }
 }
+
+const userRepository = new UsersRepository(
+  coqualityAxiosClient,
+  "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiIxIiwicm9sZSI6IlVTRVIiLCJpYXQiOjE2NzI3NTIzNTIsImV4cCI6MTY3NTM0NDM1Mn0.vY4jYVKHw9pk8LvXu8WKlse9Ncjt9qeaosFFnydN0idewco6a1ZbWP6hu1PVStqUfN-JdhBfPe-ewrDtYOaqFg"
+)
+export default userRepository
